chore(auth): Move different AWSMobileClient types to right files
