+++ AWSAuthSDK/Sources/AWSMobileClient/AWSMobileClientExtensions.swift
@@ -70,4 +70,10 @@ extension AWSMobileClient {
         }
     }
 
+    public func handleAuthResponse(_ application: UIApplication, open url: URL, sourceApplication: String?, annotation: Any) {
+        if (isCognitoAuthRegistered) {
+            AWSCognitoAuth.init(forKey: CognitoAuthRegistrationKey).application(application, open: url, options: [:])
+        }
+    }
+
 }
