+++ AWSAuthSDK/Sources/AWSMobileClient/Models/ListDevicesResult.swift
@@ -1,0 +1,27 @@
+//
+// Copyright 2017-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+// http://aws.amazon.com/apache2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
+
+import Foundation
+
+/// The result object for `list` operation in `DeviceOperations`.
+public struct ListDevicesResult {
+    public let devices: [Device]?
+    public let paginationToken: String?
+
+    internal init(devices: [Device]? = [], paginationToken: String?) {
+        self.devices = devices
+        self.paginationToken = paginationToken
+    }
+}
